requirements.txt
python-telegram-bot==12.8
requests==2.24.0
pandas==1.0.5
bs4==0.0.1
xlrd==1.2.0
typing==3.7.4.1
newrelic